requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dj-rest-auth==7.0.1
Django==5.1.2
django-allauth==65.9.0
django-cors-headers==4.5.0
django-filter==24.3
django-hosts==6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
urllib3==2.4.0
